requirements.txt
lxml,
latexfixer,
path.py,
PyYAML